# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.AddMealRequest:
    properties:
      meal_code:
        maxLength: 10
        type: string
      segment_id:
        type: string
    required:
    - meal_code
    - segment_id
    type: object
  dto.AddSSRRequest:
    properties:
      segment_id:
        type: string
      ssr_code:
        maxLength: 10
        type: string
    required:
    - segment_id
    - ssr_code
    type: object
  dto.AircraftResponse:
    properties:
      id:
        type: string
      manufacturer:
        type: string
      model:
        type: string
      seats:
        items:
          $ref: '#/definitions/dto.AircraftSeatResponse'
        type: array
      total_seats:
        type: integer
    type: object
  dto.AircraftSeatResponse:
    properties:
      id:
        type: string
      is_exit_row:
        type: boolean
      row_number:
        type: integer
      seat_class:
        $ref: '#/definitions/dto.SeatClassResponse'
      seat_letter:
        type: string
      seat_number:
        type: string
      seat_type:
        type: string
      x_position:
        type: integer
      y_position:
        type: integer
    type: object
  dto.AirportResponse:
    properties:
      city:
        type: string
      code:
        type: string
      country:
        type: string
      id:
        type: string
      name:
        type: string
      timezone:
        type: string
    type: object
...
  dto.BaggageResponse:
    properties:
      id:
        type: string
      passenger_id:
        type: string
      segment_id:
        type: string
      status:
        $ref: '#/definitions/models.BaggageStatus'
      tag_number:
        type: string
      weight:
        type: number
    type: object
...
  dto.ContactRequest:
    properties:
      email:
        maxLength: 255
        type: string
      name:
        maxLength: 255
        type: string
      phone:
        maxLength: 50
        type: string
    required:
    - email
    - name
    - phone
    type: object
...
  dto.CreateBookingRequest:
    properties:
      contact:
        $ref: '#/definitions/dto.ContactRequest'
      passengers:
        items:
          $ref: '#/definitions/dto.PassengerRequest'
        maxItems: 9
        minItems: 1
        type: array
      segments:
        items:
          $ref: '#/definitions/dto.SegmentRequest'
        maxItems: 8
        minItems: 1
        type: array
      trip_type:
        allOf:
        - $ref: '#/definitions/dto.TripType'
        enum:
        - one_way
        - round_trip
        - multi_city
    required:
    - contact
    - passengers
    - segments
    - trip_type
    type: object
...
  dto.FlightResponse:
    properties:
      aircraft:
        $ref: '#/definitions/dto.AircraftResponse'
      arrival_time:
        type: string
      available_seats:
        type: integer
      departure_time:
        type: string
      duration_minutes:
        type: integer
      flight_number:
        type: string
      id:
        type: string
      lowest_price:
        type: number
      schedule:
        $ref: '#/definitions/dto.FlightScheduleResponse'
      status:
        $ref: '#/definitions/models.FlightStatus'
    type: object
  dto.FlightScheduleResponse:
    properties:
      arrival_airport:
        $ref: '#/definitions/dto.AirportResponse'
      arrival_time:
        type: string
      departure_airport:
        $ref: '#/definitions/dto.AirportResponse'
      departure_time:
        type: string
      flight_number:
        type: string
      id:
        type: string
      operating_days:
        type: string
    type: object
...
  dto.PNRContactResponse:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
  dto.PNRListResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      pnrs:
        items:
          $ref: '#/definitions/dto.PNRResponse'
        type: array
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  dto.PNRPassengerResponse:
    properties:
      baggage:
        items:
          $ref: '#/definitions/dto.BaggageResponse'
        type: array
      birth_date:
        type: string
      first_name:
        type: string
      full_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      meals:
        items:
          $ref: '#/definitions/dto.PassengerMealResponse'
        type: array
      passenger_type:
        $ref: '#/definitions/models.PassengerType'
      passport_number:
        type: string
      seat_assignment:
        $ref: '#/definitions/dto.SeatAssignmentResponse'
      ssrs:
        items:
          $ref: '#/definitions/dto.PassengerSSRResponse'
        type: array
      ticket:
        $ref: '#/definitions/dto.TicketResponse'
    type: object
  dto.PNRResponse:
    properties:
      contact:
        $ref: '#/definitions/dto.PNRContactResponse'
      created_at:
        type: string
      id:
        type: string
      passengers:
        items:
          $ref: '#/definitions/dto.PNRPassengerResponse'
        type: array
      payments:
        items:
          $ref: '#/definitions/dto.PaymentResponse'
        type: array
      record_locator:
        type: string
      segments:
        items:
          $ref: '#/definitions/dto.PNRSegmentResponse'
        type: array
      status:
        $ref: '#/definitions/models.PNRStatus'
      total_amount:
        type: number
      trip_type:
        type: string
      ttl:
        type: string
    type: object
  dto.PNRSegmentResponse:
    properties:
      flight:
        $ref: '#/definitions/dto.FlightResponse'
      id:
        type: string
      segment_order:
        type: integer
    type: object
  dto.PassengerMealResponse:
    properties:
      id:
        type: string
      meal_code:
        type: string
      meal_name:
        type: string
      segment_id:
        type: string
    type: object
  dto.PassengerRequest:
    properties:
      birth_date:
        type: string
      first_name:
        maxLength: 100
        type: string
      last_name:
        maxLength: 100
        type: string
      passenger_type:
        allOf:
        - $ref: '#/definitions/models.PassengerType'
        enum:
        - ADT
        - CHD
        - INF
      passport_number:
        maxLength: 50
        type: string
    required:
    - first_name
    - last_name
    - passenger_type
    type: object
  dto.PassengerSSRResponse:
    properties:
      id:
        type: string
      segment_id:
        type: string
      ssr_code:
        type: string
      ssr_name:
        type: string
    type: object
  dto.PaymentResponse:
    properties:
      amount:
        type: number
      id:
        type: string
      method:
        $ref: '#/definitions/models.PaymentMethod'
      paid_at:
        type: string
      pnr_id:
        type: string
      status:
        $ref: '#/definitions/models.PaymentStatus'
    type: object
...
  dto.SeatAssignmentResponse:
    properties:
      assigned_at:
        type: string
      flight_seat_id:
        type: string
      id:
        type: string
      seat_class:
        $ref: '#/definitions/dto.SeatClassResponse'
      seat_number:
        type: string
      segment_id:
        type: string
    type: object
...
  dto.SeatClassResponse:
    properties:
      code:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  dto.SeatSelectionRequest:
    properties:
      flight_seat_id:
        type: string
      passenger_index:
        minimum: 0
        type: integer
    required:
    - flight_seat_id
    type: object
  dto.SegmentRequest:
    properties:
      flight_id:
        type: string
      seat_selections:
        items:
          $ref: '#/definitions/dto.SeatSelectionRequest'
        type: array
    required:
    - flight_id
    type: object
  dto.TicketResponse:
    properties:
      id:
        type: string
      issued_at:
        type: string
      ticket_number:
        type: string
    type: object
  dto.TripType:
    enum:
    - one_way
    - round_trip
    - multi_city
    type: string
    x-enum-varnames:
    - TripOneWay
    - TripRoundTrip
    - TripMultiCity
...
  dto.UpdateContactRequest:
    properties:
      email:
        maxLength: 255
        type: string
      name:
        maxLength: 255
        type: string
      phone:
        maxLength: 50
        type: string
    type: object
...
  models.BaggageStatus:
    enum:
    - checked_in
    - loaded
    - delivered
    - lost
    type: string
    x-enum-varnames:
    - BaggageCheckedIn
    - BaggageLoaded
    - BaggageDelivered
    - BaggageLost
...
  models.FlightStatus:
    enum:
    - scheduled
    - boarding
    - departed
    - arrived
    - cancelled
    - delayed
    type: string
    x-enum-varnames:
    - FlightStatusScheduled
    - FlightStatusBoarding
    - FlightStatusDeparted
    - FlightStatusArrived
    - FlightStatusCancelled
    - FlightStatusDelayed
  models.PNRStatus:
    enum:
    - hold
    - confirmed
    - cancelled
    - ticketed
    type: string
    x-enum-varnames:
    - PNRStatusHold
    - PNRStatusConfirmed
    - PNRStatusCancelled
    - PNRStatusTicketed
  models.PassengerType:
    enum:
    - ADT
    - CHD
    - INF
    type: string
    x-enum-comments:
      PassengerADT: Adult
      PassengerCHD: Child
      PassengerINF: Infant
    x-enum-descriptions:
    - Adult
    - Child
    - Infant
    x-enum-varnames:
    - PassengerADT
    - PassengerCHD
    - PassengerINF
  models.PaymentMethod:
    enum:
    - credit_card
    - debit_card
    - bank_transfer
    - e_wallet
    type: string
    x-enum-varnames:
    - PaymentMethodCreditCard
    - PaymentMethodDebitCard
    - PaymentMethodBankTransfer
    - PaymentMethodEWallet
  models.PaymentStatus:
    enum:
    - pending
    - success
    - failed
    - refunded
    type: string
    x-enum-varnames:
    - PaymentStatusPending
    - PaymentStatusSuccess
    - PaymentStatusFailed
    - PaymentStatusRefunded
...
  utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@airline.com
    name: API Support
  description: REST API untuk sistem Passenger Service System (PSS) maskapai penerbangan.
    Mencakup manajemen penerbangan, pemesanan tiket, check-in, bagasi, dan boarding
    pass.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Passenger Service Backend API
  version: "1.0"
paths:
...
  /bookings:
    get:
      description: Mengambil seluruh daftar PNR. Hanya admin dan super_admin.
      parameters:
      - description: 'Halaman (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Jumlah per halaman (default: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PNRListResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Daftar semua PNR
      tags:
      - Booking
    post:
      consumes:
      - application/json
      description: Membuat pemesanan tiket baru. Mendukung one_way, round_trip, dan
        multi_city. Seat akan di-lock selama 30 menit.
      parameters:
      - description: Data pemesanan
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateBookingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PNRResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Penerbangan atau kursi tidak ditemukan
          schema:
            $ref: '#/definitions/utils.Response'
        "409":
          description: Kursi sudah dipesan
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Buat pemesanan (PNR)
      tags:
      - Booking
  /bookings/{id}:
    delete:
      description: Membatalkan PNR dan melepaskan kursi yang di-lock. PNR yang sudah
        ticketed tidak bisa dibatalkan.
      parameters:
      - description: PNR UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "422":
          description: PNR sudah dibatalkan atau sudah ticketed
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Batalkan pemesanan
      tags:
      - Booking
    get:
      description: Mengambil detail PNR lengkap termasuk penumpang, segmen, kursi,
        dan pembayaran.
      parameters:
      - description: PNR UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PNRResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Detail pemesanan berdasarkan ID
      tags:
      - Booking
  /bookings/{id}/contact:
    put:
      consumes:
      - application/json
      description: Memperbarui informasi kontak (nama, email, telepon) pada sebuah
        PNR.
      parameters:
      - description: PNR UUID
        in: path
        name: id
        required: true
        type: string
      - description: Data kontak baru
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateContactRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PNRContactResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Update kontak pemesanan
      tags:
      - Booking
  /bookings/locator/{locator}:
    get:
      description: 'Mengambil detail PNR menggunakan 6-karakter record locator (contoh:
        ABC123).'
      parameters:
      - description: Record locator (6 karakter)
        example: ABC123
        in: path
        name: locator
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PNRResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Cari pemesanan berdasarkan kode booking
      tags:
      - Booking
  /bookings/meal/{mealID}:
    delete:
      description: Menghapus pilihan meal penumpang.
      parameters:
      - description: Meal UUID
        in: path
        name: mealID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Hapus pilihan meal
      tags:
      - Booking
  /bookings/passengers/{passengerID}/meal:
    post:
      consumes:
      - application/json
      description: Menambahkan pilihan meal ke penumpang untuk segment tertentu.
      parameters:
      - description: Passenger UUID
        in: path
        name: passengerID
        required: true
        type: string
      - description: Data meal
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AddMealRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PassengerMealResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Tambah pilihan meal
      tags:
      - Booking
  /bookings/passengers/{passengerID}/ssr:
    post:
      consumes:
      - application/json
      description: Menambahkan Special Service Request (kursi roda, meal khusus, dll)
        ke penumpang.
      parameters:
      - description: Passenger UUID
        in: path
        name: passengerID
        required: true
        type: string
      - description: Data SSR
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AddSSRRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PassengerSSRResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Tambah SSR ke penumpang
      tags:
      - Booking
  /bookings/ssr/{ssrID}:
    delete:
      description: Menghapus SSR yang sudah ditambahkan ke penumpang.
      parameters:
      - description: SSR UUID
        in: path
        name: ssrID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Hapus SSR penumpang
      tags:
      - Booking
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"